# components/backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/initiate-registration": {
      "post": {
        "description": "Инициализация регистрации",
        "operationId": "initiateRegistration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateRegistrationInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная инициализация регистрации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateRegistrationInputDTO"
                }
              }
            }
          }
        },
        "summary": "initiateRegistration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/complete-registration": {
      "post": {
        "description": "Завершение регистрации",
        "operationId": "completeRegistration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRegistrationInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное завершение регистрации",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteLoginResponseDTO"
                }
              }
            }
          }
        },
        "summary": "completeRegistration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/initiate-login": {
      "post": {
        "description": "Инициализация входа",
        "operationId": "initiateLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateLoginInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная инициализация входа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateLoginInputDTO"
                }
              }
            }
          }
        },
        "summary": "initiateLogin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/complete-login": {
      "post": {
        "description": "Завершение входа",
        "operationId": "completeLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteLoginInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное завершение входа",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteLoginResponseDTO"
                }
              }
            }
          }
        },
        "summary": "completeLogin",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh-token": {
      "post": {
        "description": "Обновление токена",
        "operationId": "refreshAccessToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное обновление токена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteLoginResponseDTO"
                }
              }
            }
          }
        },
        "summary": "refreshAccessToken",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Выход",
        "operationId": "logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный выход"
          }
        },
        "summary": "logout",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/access/prepare-share-data": {
      "post": {
        "description": "Подготовка данных для общего доступа",
        "operationId": "prepareShareData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareShareDataInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная подготовка данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CoopInfoResponseDTO"
                }
              }
            }
          }
        },
        "summary": "prepareShareData",
        "tags": [
          "Access"
        ]
      }
    },
    "/access/share-data": {
      "post": {
        "description": "Передача данных",
        "operationId": "shareData",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareDataDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешная передача данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareDataResponseDTO"
                }
              }
            }
          }
        },
        "summary": "shareData",
        "tags": [
          "Access"
        ]
      }
    },
    "/access/exchange-ticket": {
      "post": {
        "description": "Обмен тикета на JWT",
        "operationId": "exchangeTicketForJwt",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeTicketInputDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешный обмен тикета на JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExchangeTicketResponseDTO"
                }
              }
            }
          }
        },
        "summary": "exchangeTicketForJwt",
        "tags": [
          "Access"
        ]
      }
    },
    "/access/get-encrypted-data/{username}/{coopname}": {
      "get": {
        "description": "Получение зашифрованных данных",
        "operationId": "getEncryptedData",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "coopname",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное получение зашифрованных данных",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EncryptedDataResponseDTO"
                }
              }
            }
          }
        },
        "summary": "getEncryptedData",
        "tags": [
          "Access"
        ]
      }
    }
  },
  "info": {
    "title": "CARD.COOP",
    "description": "API для авторизации и управления пользователями",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3015"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "InitiateRegistrationInputDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CompleteRegistrationInputDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "hash_key": {
            "type": "string",
            "description": "Хэш-ключ для завершения регистрации",
            "example": "secureHash123"
          },
          "uuid": {
            "type": "string",
            "description": "Уникальный идентификатор запроса",
            "example": "uuid-123e4567-e89b-12d3-a456-426614174000"
          },
          "salt": {
            "type": "string",
            "description": "Серверная соль для защиты данных",
            "example": "random-server-salt-example"
          }
        },
        "required": [
          "email",
          "hash_key",
          "uuid",
          "salt"
        ]
      },
      "CompleteLoginResponseDTO": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Токен доступа для авторизации",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refresh_token": {
            "type": "string",
            "description": "Токен для обновления доступа",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
      "InitiateLoginInputDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CompleteLoginInputDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email пользователя",
            "example": "user@example.com"
          },
          "hash_key": {
            "type": "string",
            "description": "Хэш-ключ для завершения входа",
            "example": "secureHash123"
          },
          "uuid": {
            "type": "string",
            "description": "Уникальный идентификатор запроса",
            "example": "uuid-123e4567-e89b-12d3-a456-426614174000"
          }
        },
        "required": [
          "email",
          "hash_key",
          "uuid"
        ]
      },
      "RefreshTokenInputDTO": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh-токен для обновления токена доступа",
            "example": "refreshtoken123"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "LogoutInputDTO": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Access-токен для удаления",
            "example": "accesstoken123"
          },
          "refresh_token": {
            "type": "string",
            "description": "Refresh-токен для удаления",
            "example": "refreshtoken123"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ]
      },
...
      "PrepareShareDataInputDTO": {
        "type": "object",
        "properties": {
          "coopname": {
            "type": "string",
            "description": "Название кооператива",
            "example": "CoopNameXYZ"
          }
        },
        "required": [
          "coopname"
        ]
      },
      "CoopInfoResponseDTO": {
        "type": "object",
        "properties": {
          "coopname": {
            "type": "string",
            "description": "Имя аккаунта кооператива",
            "example": "voskhod"
          },
          "public_key": {
            "type": "string",
            "description": "Публичный ключ кооператива",
            "example": "EOS123456789abcdef"
          },
          "announce": {
            "type": "string",
            "description": "Публичное наименование кооператива",
            "example": "Потребительский Кооператив \"РОМАШКА\""
          }
        },
        "required": [
          "coopname",
          "public_key",
          "announce"
        ]
      },
      "AccessMetaInputDTO": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": "string",
            "description": "Идентификатор устройства пользователя",
            "example": "device12345"
          },
          "device_type": {
            "type": "string",
            "description": "Тип устройства пользователя",
            "example": "desktop"
          },
          "operating_system": {
            "type": "string",
            "description": "Операционная система устройства",
            "example": "Windows 10"
          },
          "browser": {
            "type": "string",
            "description": "Браузер, через который был предоставлен доступ",
            "example": "Chrome 114"
          }
        }
      },
      "ShareDataDTO": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя аккаунта пользователя",
            "example": "username"
          },
          "coopname": {
            "type": "string",
            "description": "Имя аккаунта кооператива",
            "example": "voskhod"
          },
          "encrypted_data": {
            "type": "string",
            "description": "Зашифрованные данные для передачи",
            "example": "encryptedData123456abcdef"
          },
          "public_key": {
            "type": "string",
            "description": "Публичный ключ кооператива необходимый для дешифровки",
            "example": "EOSPublicKey123456abcdef"
          },
          "meta": {
            "description": "Не обязательные данные для идентификации устройства пользователя",
            "allOf": [
              {
                "$ref": "#/components/schemas/AccessMetaInputDTO"
              }
            ]
          }
        },
        "required": [
          "username",
          "coopname",
          "encrypted_data",
          "public_key"
        ]
      },
      "ShareDataResponseDTO": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "description": "Билет для обмена данными",
            "example": "exchangeTicket12345"
          }
        },
        "required": [
          "ticket"
        ]
      },
      "ExchangeTicketInputDTO": {
        "type": "object",
        "properties": {
          "ticket": {
            "type": "string",
            "description": "Билет для обмена кооперативной информацией",
            "example": "exchangeTicket12345"
          }
        },
        "required": [
          "ticket"
        ]
      },
      "ExchangeTicketResponseDTO": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "JWT токен доступа для взаимодействия с кооперативом",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "access_token"
        ]
      },
      "Function": {
        "type": "object",
        "properties": {}
      },
      "EncryptedDataResponseDTO": {
        "type": "object",
        "properties": {
          "encrypted_data": {
            "type": "string",
            "description": "Зашифрованные данные",
            "example": "encrypted_string_example"
          },
          "coopname": {
            "type": "string",
            "description": "Имя кооператива",
            "example": "voskhod"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "user123"
          },
          "public_key": {
            "type": "string",
            "description": "Публичный ключ",
            "example": "public_key_example"
          },
          "meta": {
            "description": "Метаданные доступа",
            "allOf": [
              {
                "$ref": "#/components/schemas/Function"
              }
            ]
          }
        },
        "required": [
          "encrypted_data",
          "coopname",
          "username",
          "public_key",
          "meta"
        ]
      }
    }
  }
}